+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/GDXRoot.java
@@ -11,12 +11,13 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * Updated asset version, 2/6/2021
  */
- package edu.cornell.gdiac.physics;
+ package edu.cornell.gdiac.honeyHeistCode;
 
 import com.badlogic.gdx.*;
+import edu.cornell.gdiac.honeyHeistCode.controllers.LevelController;
+import edu.cornell.gdiac.honeyHeistCode.controllers.LoadingMode;
 import edu.cornell.gdiac.util.*;
 import edu.cornell.gdiac.assets.*;;
-import edu.cornell.gdiac.physics.platform.*;
 
 /**
  * Root class for a LibGDX.
