+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/controllers/AIController.java
@@ -6,12 +6,14 @@ will have an AI Controller assigned to them.
 
 It is dependent on Level Model in order for the character to get information about the level.
  */
-package edu.cornell.gdiac.physics.platform;
+package edu.cornell.gdiac.honeyHeistCode.controllers;
 
 import com.badlogic.gdx.math.Vector2;
-import edu.cornell.gdiac.physics.obstacle.ComplexObstacle;
-import edu.cornell.gdiac.physics.obstacle.Obstacle;
-import edu.cornell.gdiac.physics.obstacle.PolygonObstacle;
+import edu.cornell.gdiac.honeyHeistCode.models.ChaserBeeModel;
+import edu.cornell.gdiac.honeyHeistCode.models.LevelModel;
+import edu.cornell.gdiac.honeyHeistCode.models.PlatformModel;
+import edu.cornell.gdiac.honeyHeistCode.obstacle.Obstacle;
+
 import java.util.Random;
 
 
@@ -68,12 +70,12 @@ public class AIController {
 	}
 
     private void updateLineToPlayer() {
-        lineToPlayer.set(levelModel.getAvatar().getPosition());
+        lineToPlayer.set(levelModel.getPlayer().getPosition());
         lineToPlayer.sub(controlledEnemy.getPosition());
     }
 
 	private void updateFSMState() {
-		float distanceToPlayer = controlledEnemy.getPosition().dst(levelModel.getAvatar().getPosition());
+		float distanceToPlayer = controlledEnemy.getPosition().dst(levelModel.getPlayer().getPosition());
 		switch (this.state) {
 			case WANDER:
 				if (distanceToPlayer < CHASE_RADIUS) {
