+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/controllers/InputController.java
@@ -11,7 +11,7 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * LibGDX version, 2/6/2015
  */
-package edu.cornell.gdiac.physics;
+package edu.cornell.gdiac.honeyHeistCode.controllers;
 
 import com.badlogic.gdx.*;
 import com.badlogic.gdx.math.*;
