+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/controllers/LevelController.java
@@ -8,7 +8,7 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * Updated asset version, 2/6/2021
  */
-package edu.cornell.gdiac.physics.platform;
+package edu.cornell.gdiac.honeyHeistCode.controllers;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
@@ -19,10 +19,10 @@ import com.badlogic.gdx.utils.JsonValue;
 import com.badlogic.gdx.utils.ObjectSet;
 import edu.cornell.gdiac.assets.AssetDirectory;
 import edu.cornell.gdiac.audio.SoundBuffer;
-import edu.cornell.gdiac.physics.InputController;
-import edu.cornell.gdiac.physics.WorldController;
-import edu.cornell.gdiac.physics.obstacle.BoxObstacle;
-import edu.cornell.gdiac.physics.obstacle.Obstacle;
+import edu.cornell.gdiac.honeyHeistCode.WorldController;
+import edu.cornell.gdiac.honeyHeistCode.models.*;
+import edu.cornell.gdiac.honeyHeistCode.obstacle.BoxObstacle;
+import edu.cornell.gdiac.honeyHeistCode.obstacle.Obstacle;
 
 /**
  * Gameplay specific controller for the platformer game.
@@ -178,11 +178,11 @@ public class LevelController extends WorldController implements ContactListener 
         // Create player (ant)
         dwidth = avatarTexture.getRegionWidth() / scale.x;
         dheight = avatarTexture.getRegionHeight() / scale.y;
-        AntModel avatar = new AntModel(constants.get("player"), dwidth, dheight);
+        Player avatar = new Player(constants.get("player"), dwidth, dheight);
         avatar.setDrawScale(scale);
         avatar.setTexture(avatarTexture);
         addObject(avatar);
-        level.setAvatar(avatar);
+        level.setPlayer(avatar);
 
         // Create one chaser bee
         dwidth = chaserBeeTexture.getRegionWidth() / scale.x;
@@ -217,7 +217,7 @@ public class LevelController extends WorldController implements ContactListener 
      */
     public void rotate(boolean isClockwise, boolean platformNotRotating){
         PlatformModel platforms = level.getPlatforms();
-        AntModel avatar = level.getAvatar();
+        Player avatar = level.getPlayer();
         Array<AbstractBeeModel> bees = level.getBees();
         Vector2 origin = level.getOrigin();
 
@@ -257,7 +257,7 @@ public class LevelController extends WorldController implements ContactListener 
      * @param direction -1 = left, 1 = right, 0 = still
      */
     public void moveAnt(float direction) {
-        level.getAvatar().setMovement(direction * level.getAvatar().getForce());
+        level.getPlayer().setMovement(direction * level.getPlayer().getForce());
     }
 
     /**
@@ -284,7 +284,7 @@ public class LevelController extends WorldController implements ContactListener 
             return false;
         }
 
-        if (!isFailure() && level.getAvatar().getY() < -1) {
+        if (!isFailure() && level.getPlayer().getY() < -1) {
             setFailure(true);
             return false;
         }
@@ -306,7 +306,7 @@ public class LevelController extends WorldController implements ContactListener 
         // Process actions in object model
         moveAnt(InputController.getInstance().getHorizontal());
 
-        level.getAvatar().applyForce();
+        level.getPlayer().applyForce();
 
         // Process AI action
         // 1. Loop over all chaser bee,
@@ -345,7 +345,7 @@ public class LevelController extends WorldController implements ContactListener 
         Object fd1 = fix1.getUserData();
         Object fd2 = fix2.getUserData();
 
-        AntModel avatar = level.getAvatar();
+        Player avatar = level.getPlayer();
         Array<AbstractBeeModel> bees = level.getBees();
         BoxObstacle goalDoor = level.getGoalDoor();
 
@@ -398,7 +398,7 @@ public class LevelController extends WorldController implements ContactListener 
         Object bd1 = body1.getUserData();
         Object bd2 = body2.getUserData();
 
-        AntModel avatar = level.getAvatar();
+        Player avatar = level.getPlayer();
         Array<AbstractBeeModel> bees = level.getBees();
 
         if ((avatar.getSensorName().equals(fd2) && avatar != bd1) ||
