+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/models/ChaserBeeModel.java
@@ -1,4 +1,4 @@
-package edu.cornell.gdiac.physics.platform;
+package edu.cornell.gdiac.honeyHeistCode.models;
 
 import com.badlogic.gdx.utils.JsonValue;
