+++ HoneyHeist/core/src/edu/cornell/gdiac/honeyHeistCode/models/LevelModel.java
@@ -8,11 +8,11 @@
  * Based on original PhysicsDemo Lab by Don Holden, 2007
  * Updated asset version, 2/6/2021
  */
-package edu.cornell.gdiac.physics.platform;
+package edu.cornell.gdiac.honeyHeistCode.models;
 
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
-import edu.cornell.gdiac.physics.obstacle.BoxObstacle;
+import edu.cornell.gdiac.honeyHeistCode.obstacle.BoxObstacle;
 
 /**
  * A class holding all game objects contained within a level,
@@ -25,7 +25,7 @@ public class LevelModel {
     /**
      * Reference to the character avatar
      */
-    private AntModel avatar;
+    private Player player;
     /**
      * Reference to the list of bees
      */
@@ -58,11 +58,11 @@ public class LevelModel {
 
     public void setPlatforms(PlatformModel platforms) { this.platforms = platforms; }
 
-    public AntModel getAvatar() {
-        return avatar;
+    public Player getPlayer() {
+        return player;
     }
 
-    public void setAvatar(AntModel avatar) { this.avatar = avatar; }
+    public void setPlayer(Player player) { this.player = player; }
 
     public Array<AbstractBeeModel> getBees() {return bees;}
